+++ tiger-testenv-mgr/src/test/java/de/gematik/test/tiger/testenvmgr/servers/log/TigerServerLogManagerTest.java
@@ -29,7 +29,7 @@ class TigerServerLogManagerTest {
     @AfterEach
     void setup() {
         TigerGlobalConfiguration.reset();
-        Path.of("target", "serverLogs", "test.log").toFile().delete();
+        Path.of("target", "serverLogs").toFile().deleteOnExit();
     }
 
     @SystemStub
@@ -42,7 +42,7 @@ class TigerServerLogManagerTest {
         String serverID = "ExternalJar-001";
         final CfgServer configuration = new CfgServer();
         configuration.setExternalJarOptions(new CfgExternalJarOptions());
-        configuration.getExternalJarOptions().setActivateLogs(false);
+        configuration.getExternalJarOptions().setActivateLogs(true);
         configuration.setType(ExternalJarServer.class.getAnnotation(TigerServerType.class));
         configuration.setLogFile(logFile);
         ExternalJarServer server = ExternalJarServer.builder().serverId(serverID).configuration(configuration).build();
@@ -66,7 +66,7 @@ class TigerServerLogManagerTest {
                   - local:winstone.jar
                 healthcheckUrl: http://127.0.0.1:${free.port.0}
                 healthcheckReturnCode: 200
-                logFile: target/serverLogs/test.log
+                logFile: target/serverLogs/test1.log
                 externalJarOptions:
                   workingDir: target
                   arguments:
@@ -75,7 +75,7 @@ class TigerServerLogManagerTest {
             """)
     @Test
     void testCheckAddAppendersEnabledLog_OK() throws Exception {
-        assertThat(new File("target/serverLogs/test.log")).content()
+        assertThat(new File("target/serverLogs/test1.log")).content()
                 .contains("Winstone Servlet Engine ");
     }
 
@@ -88,7 +88,7 @@ class TigerServerLogManagerTest {
                   - local:winstone.jar
                 healthcheckUrl: http://127.0.0.1:${free.port.0}
                 healthcheckReturnCode: 200
-                logFile: target/serverLogs/test.log
+                logFile: target/serverLogs/test2.log
                 externalJarOptions:
                   activateLogs: false
                   workingDir: target
@@ -97,8 +97,54 @@ class TigerServerLogManagerTest {
                     - "--webroot=."
             """)
     @Test
-    void testCheckAddAppendersDisabledLog_OK() throws Exception {
-        assertThat(new File("target/serverLogs/test.log")).content()
-                .contains("Winstone Servlet Engine ");
+    void testCheckAddAppendersDisabledLog_OK() {
+        assertThat(new File("target/serverLogs/test2.log")).doesNotExist();
+    }
+
+
+    @TigerTest(tigerYaml = """
+            localProxyActive: false
+            servers:
+              externalJarServer:
+                type: externalJar
+                source:
+                  - local:winstone.jar
+                healthcheckUrl: http://127.0.0.1:${free.port.0}
+                healthcheckReturnCode: 200
+                logFile: target/serverLogs/test3.log
+                externalJarOptions:
+                  activateWorkflowLogs: false
+                  workingDir: target
+                  arguments:
+                    - "--httpPort=${free.port.0}"
+                    - "--webroot=."
+            """)
+    @Test
+    void testCheckAddAppendersDisabledOnlyWorkflowUiLog_OK() {
+        assertThat(new File("target/serverLogs/test3.log")).content()
+            .contains("Winstone Servlet Engine ");
+    }
+
+    @TigerTest(tigerYaml = """
+            localProxyActive: false
+            servers:
+              externalJarServer:
+                type: externalJar
+                source:
+                  - local:winstone.jar
+                healthcheckUrl: http://127.0.0.1:${free.port.0}
+                healthcheckReturnCode: 200
+                logFile: target/serverLogs/test4.log
+                externalJarOptions:
+                  activateLogs: false
+                  activateWorkflowLogs: true
+                  workingDir: target
+                  arguments:
+                    - "--httpPort=${free.port.0}"
+                    - "--webroot=."
+            """)
+    @Test
+    void testCheckAddAppendersDisabledAllButOnlyWorkflowUiLog_OK() {
+        assertThat(new File("target/serverLogs/test4.log")).doesNotExist();
     }
 }