+++ ReleaseNotes.md
@@ -12,6 +12,7 @@
 # Release 2.1.8
 
 * TGR-1048: Resolved the problems with steps not being found in IntelliJ's Gherkin plugin. Also they were executed, one was not able to look them up with auto complete and if entered manually the step was marked as undefined
+* TGR-1045: With "activateLogs: false" in externalJarOptions for servers deactivates the logs completely, with "activateWorkflowLogs: false" deactivates sending logs to the workflow UI only
 
 -------
 * Serenity BDD 3.6.23